fix warning in Listing 1.3

+++ LitvinenkoNA_C++ProgrammingWin32ApiApp/Listings/Chapter1/Listing_1_3/Listing_1_3.cpp
@@ -137,7 +137,7 @@ LRESULT CALLBACK WndProc(HWND hWnd, UINT message, WPARAM wParam, LPARAM lParam)
         break;
     case WM_PAINT:
         hdc = BeginPaint(hWnd, &ps);
-        TextOut(hdc, 0, 0, str.data(), str.size());
+        TextOut(hdc, 0, 0, str.data(), static_cast<int>(str.size()));
         EndPaint(hWnd, &ps);
         break;
     case WM_DESTROY: PostQuitMessage(0); break;
